Add files via upload
The main objective to develop a Hospital management system in Java is to manage the
hospital activity. It uses Java as core technology and Mysql as backend to manage the data
records.The application is following the MVC architecture with maven tool.This application helps to
manage the activity of a hospital like Patient management, Doctor management, Manage
appointments. Manage the records of patients.